# api/openapi.yaml
openapi: 3.1.0
info:
  title: Findrail local search API
  version: 0.1.0-alpha.1
  description: The API binds to loopback and returns indexed text and local paths,
    which clients must treat as private. `serve` and `demo` expose read-only routes.
    Only `start` enables the source mutation routes below. Mutations require an exact
    Origin for the actual bound authority, `X-Findrail-Token` from the no-store
    session endpoint, non-cross-site Fetch Metadata, and strict JSON up to 16 KiB.
    This is browser-origin protection, not authentication against same-user processes.
servers:
- url: http://127.0.0.1:7766
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Check that the HTTP process is responding
      responses:
        '200':
          description: HTTP process is available; not a full index integrity check
          content:
            application/json:
              schema:
                type: object
                required:
                - status
                properties:
                  status:
                    type: string
                    const: ok
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/search:
    get:
      operationId: search
      summary: Search indexed documents in Literal or opt-in Advanced mode
      parameters:
      - name: q
        in: query
        required: true
        description: Literal mode tokenizes punctuation as separators. Advanced mode uses the restricted grammar documented in docs/search-query.md; user input is compiled to safe FTS expressions, not passed through as SQL or raw MATCH syntax.
        schema:
          type: string
          minLength: 1
          maxLength: 256
      - name: source
        in: query
        description: Optional exact source ID; an unknown ID yields no matches
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: mode
        in: query
        description: Literal is the compatible default; Advanced enables the restricted query grammar in docs/search-query.md.
        schema: {type: string, enum: ['', literal, advanced], default: literal}
      - name: format
        in: query
        description: Server-side media type filter.
        schema: {type: string, enum: ['', all, text, pdf, docx], default: all}
      - name: path_prefix
        in: query
        description: Relative indexed slash path; literal prefix with directory boundary.
        schema: {type: string, maxLength: 512}
      - name: title_contains
        in: query
        description: Literal, case-sensitive title substring.
        schema: {type: string, maxLength: 128}
      responses:
        '200':
          description: Ranked matches and total match count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
  /api/v1/sources:
    get:
      operationId: listSources
      summary: List explicitly indexed sources and freshness
      responses:
        '200':
          description: Source inventory
          content:
            application/json:
              schema:
                type: object
                required:
                - sources
                properties:
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/Source'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
  /api/v1/sources/{id}/report:
    get:
      operationId: getSourceIndexingReport
      summary: Read the latest committed source indexing report without triggering a scan
      parameters:
      - {$ref: '#/components/parameters/SourceID'}
      - name: include_paths
        in: query
        description: Explicitly include at most 200 safe relative-path examples (at most 64 KiB serialized report).
        schema: {type: boolean, default: false}
      responses:
        '200': {description: Committed report, or explicit unavailable/frozen-origin status}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {description: Source does not exist}
        '500': {$ref: '#/components/responses/Unavailable'}
    post:
      operationId: addSource
      summary: Queue an explicitly selected folder or public GitHub source (start only)
      parameters:
      - $ref: '#/components/parameters/ManagementToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [type]
              properties:
                type: {type: string, enum: [folder, github]}
                path: {type: string, maxLength: 4096}
                repository: {type: string, maxLength: 300}
                ref: {type: string, maxLength: 1024}
                subdirectory: {type: string, maxLength: 1024}
                max_docx_bytes:
                  type: integer
                  minimum: 0
                  maximum: 16777216
                  default: 8388608
                  description: Folder DOCX input budget; 0 disables DOCX.
      responses:
        '202': {description: Job accepted}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {description: Duplicate source or busy target}
        '413': {description: Request exceeds 16 KiB}
        '429': {description: Job queue is full}
        '503': {description: Server is stopping}
  /api/v1/capabilities:
    get:
      operationId: getCapabilities
      summary: Discover whether this composition enables source management
      responses:
        '200': {description: Management capability; false for serve and demo}
  /api/v1/session:
    get:
      operationId: getManagementSession
      summary: Get the process-scoped in-memory mutation capability (start only)
      responses:
        '200': {description: No-store session token}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/sources/{id}/refresh:
    post:
      operationId: refreshSource
      summary: Queue a manual refresh using the stored source configuration (start only)
      parameters:
      - {$ref: '#/components/parameters/SourceID'}
      - {$ref: '#/components/parameters/ManagementToken'}
      description: Send an empty body with Content-Type application/json.
      responses:
        '202': {description: Refresh job accepted}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Unknown source}
        '409': {description: Source already has an operation}
        '429': {description: Job queue is full}
  /api/v1/sources/{id}:
    delete:
      operationId: removeSource
      summary: Queue confirmed logical removal from the local index (start only)
      parameters:
      - {$ref: '#/components/parameters/SourceID'}
      - {$ref: '#/components/parameters/ManagementToken'}
      description: Send an empty body with Content-Type application/json. Original files are untouched.
      responses:
        '202': {description: Removal job accepted; poll its job ID until terminal}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Unknown source}
        '409': {description: Removal already queued}
        '429': {description: Job queue is full}
  /api/v1/sources/{id}/configure:
    post:
      operationId: configureFolder
      summary: Atomically update DOCX policy and re-index a filesystem source (start only)
      parameters:
      - {$ref: '#/components/parameters/SourceID'}
      - {$ref: '#/components/parameters/ManagementToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [max_docx_bytes]
              properties:
                max_docx_bytes: {type: integer, minimum: 0, maximum: 16777216}
      responses:
        '202': {description: Configure job accepted; poll its job ID until terminal}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Unknown source}
        '409': {description: Source changed or another operation is active}
        '429': {description: Job queue is full}
  /api/v1/jobs:
    get:
      operationId: listJobs
      summary: List bounded process-memory job status (start only)
      responses:
        '200': {description: At most one hundred completed jobs plus active jobs}
  /api/v1/jobs/{id}:
    get:
      operationId: getJob
      summary: Read one process-memory job
      parameters:
      - {$ref: '#/components/parameters/JobID'}
      responses:
        '200': {description: Job state}
        '404': {description: Unknown or expired job}
  /api/v1/jobs/{id}/cancel:
    post:
      operationId: cancelJob
      summary: Cancel a queued/running job (start only)
      parameters:
      - {$ref: '#/components/parameters/JobID'}
      - {$ref: '#/components/parameters/ManagementToken'}
      description: Send an empty body with Content-Type application/json. Running cancellation returns only after operation cleanup when polled.
      responses:
        '204': {description: Cancellation accepted or job already terminal}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Unknown or expired job}
  /api/v1/documents/{id}:
    get:
      operationId: getEvidence
      summary: Preview an indexed document or PDF page
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: PDF page; omitted means page 1. Page 0 requests a PDF title-only result without a page claim. Text documents must omit it.
        schema:
          type: integer
          minimum: 0
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
        '404':
          description: Document no longer exists in the indexed snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Bounded plain text from the indexed snapshot, never an arbitrary
            file read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evidence'
  /api/v1/sync:
    get:
      operationId: getSyncStatus
      summary: Current in-memory source refresh health
      responses:
        '200':
          description: Refresh mode and worker status; rebuilt after restart
          content:
            application/json:
              schema:
                type: object
                required:
                - enabled
                - sources
                properties:
                  enabled:
                    type: boolean
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncStatus'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  parameters:
    ManagementToken:
      name: X-Findrail-Token
      in: header
      required: true
      description: In-memory capability returned by GET /api/v1/session; never put it in a URL.
      schema: {type: string, minLength: 32, maxLength: 64}
    SourceID:
      name: id
      in: path
      required: true
      schema: {type: string}
    JobID:
      name: id
      in: path
      required: true
      schema: {type: string, minLength: 32, maxLength: 32}
  responses:
    BadRequest:
      description: Invalid query, mode, search filter, result limit or PDF page
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalidAdvancedQuery:
              summary: An OR branch has no positive term
              value: {error: "Each OR branch needs a positive term.", code: missing_positive_term, position: 9}
    Forbidden:
      description: Non-local Host or a browser Origin that differs from the request
        Host
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unavailable:
      description: Index operation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        code:
          type: string
          enum: [syntax, missing_positive_term, query_too_large, invalid_mode, invalid_filter]
          description: Stable search validation category; present for search validation errors.
        position:
          type: integer
          minimum: 0
          description: Zero-based Unicode code-point offset when the error points into the query.
    SearchResponse:
      type: object
      required:
      - query
      - total
      - results
      properties:
        query:
          type: string
        total:
          type: integer
          minimum: 0
        results:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/SearchResult'
    SearchResult:
      type: object
      required:
      - id
      - title
      - uri
      - path
      - source_id
      - source_name
      - source_kind
      - snippet
      - score
      - media_type
      properties:
        id: &id001
          type: string
        title: &id002
          type: string
        uri: &id003
          type: string
          description: Original location, not an HTTP document download endpoint
        path: &id004
          type: string
          description: Relative original path using forward slashes
        source_id: &id005
          type: string
        source_name: &id006
          type: string
        source_kind:
          type: string
        snippet:
          type: string
          description: Plain-text excerpt; square brackets mark matching terms
        score:
          type: number
          description: Negative SQLite BM25 value; higher ranks first, not a probability
        media_type: &id007
          type: string
        page: &id008
          type: integer
          minimum: 0
          description: Positive-evidence PDF page, or 0 when Advanced matches only the title.
        page_count: &id009
          type: integer
          minimum: 1
    Source:
      type: object
      required:
      - id
      - kind
      - name
      - root
      - documents
      - last_indexed_at
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [filesystem, github, archive]
        name:
          type: string
        root:
          type: string
          description: Canonical absolute local directory
        documents:
          type: integer
          minimum: 0
        last_indexed_at:
          type: string
          format: date-time
        max_text_bytes:
          type: integer
        max_pdf_bytes:
          type: integer
          description: 0 disables PDF extraction for this source
        max_docx_bytes:
          type: integer
          description: Folder DOCX input limit in bytes; 0 disables DOCX (legacy sources default to 0).
        github:
          type: object
          description: Present only for public GitHub snapshots
          properties:
            owner: {type: string}
            repo: {type: string}
            ref_mode: {type: string, enum: [default, ref, commit]}
            ref_value: {type: string}
            selected_path: {type: string}
            sha: {type: string, pattern: '^[0-9a-f]{40}$'}
            last_error: {type: string, maxLength: 512}
        indexing_report:
          $ref: '#/components/schemas/IndexingReportSummary'
    IndexingReportSummary:
      type: object
      required: [available]
      properties:
        available: {type: boolean}
        availability: {type: string, enum: [not_yet_available, frozen_origin_report_not_in_portable_v1]}
        report_id: {type: string}
        operation: {type: string, enum: [initial_index, refresh, configure, github_publication]}
        finished_at: {type: string, format: date-time}
        indexed_documents: {type: integer, minimum: 0}
        skipped_files: {type: integer, minimum: 0}
        pruned_directories: {type: integer, minimum: 0}
        coverage: {type: string}
        reasons:
          type: array
          maxItems: 15
          items:
            type: object
            required: [code, count, unit]
            properties:
              code: {type: string}
              count: {type: integer, minimum: 0}
              unit: {type: string, enum: [file, directory, entry]}
    Evidence:
      type: object
      required:
      - id
      - title
      - uri
      - path
      - source_id
      - source_name
      - media_type
      - content_hash
      - modified_at
      - text
      - truncated
      properties:
        id: *id001
        title: *id002
        uri: *id003
        path: *id004
        source_id: *id005
        source_name: *id006
        media_type: *id007
        page: *id008
        page_count: *id009
        content_hash:
          type: string
        modified_at:
          type: string
          format: date-time
        text:
          type: string
          maxLength: 65536
        truncated:
          type: boolean
    SyncStatus:
      type: object
      required:
      - source_id
      - state
      - mode
      - skipped
      properties:
        source_id:
          type: string
        state:
          type: string
          enum:
          - starting
          - indexing
          - idle
          - error
        mode:
          type: string
          enum:
          - events
          - polling
        last_started_at:
          type: string
          format: date-time
        last_success_at:
          type: string
          format: date-time
        last_error:
          type: string
        watch_warning:
          type: string
        skipped:
          type: integer
          minimum: 0
        skipped_pdf:
          type: integer
          minimum: 0